+++ src/main/java/com/box/sdk/BoxAPIConnection.java
@@ -84,7 +84,7 @@ public class BoxAPIConnection {
     private static final String BOX_NOTIFICATIONS_HEADER = "Box-Notifications";
 
     private static final String JAVA_VERSION = System.getProperty("java.version");
-    private static final String SDK_VERSION = "3.8.0";
+    private static final String SDK_VERSION = "3.8.2";
 
     /**
      * The amount of buffer time, in milliseconds, to use when determining if an access token should be refreshed. For
@@ -822,17 +822,20 @@ public class BoxAPIConnection {
     public void restore(String state) {
         JsonObject json = Json.parse(state).asObject();
         String accessToken = json.get("accessToken").asString();
-        String refreshToken = Optional.ofNullable(json.get("refreshToken"))
-            .map(v -> v.isNull() ? null : v.asString())
-            .orElse(null);
+        String refreshToken = getKeyValueOrDefault(json, "refreshToken", null);
         long lastRefresh = json.get("lastRefresh").asLong();
         long expires = json.get("expires").asLong();
         String userAgent = json.get("userAgent").asString();
         String tokenURL = getKeyValueOrDefault(json, "tokenURL", null);
         String revokeURL = getKeyValueOrDefault(json, "revokeURL", null);
-        String baseURL = json.get("baseURL").asString();
-        String baseUploadURL = json.get("baseUploadURL").asString();
-        String authorizationURL = getKeyValueOrDefault(json, "authorizationURL", DEFAULT_BASE_AUTHORIZATION_URL);
+        String baseURL = adoptBaseUrlWhenLoadingFromOldVersion(
+            getKeyValueOrDefault(json, "baseURL", DEFAULT_BASE_URL)
+        );
+        String baseUploadURL = adoptUploadBaseUrlWhenLoadingFromOldVersion(
+            getKeyValueOrDefault(json, "baseUploadURL", DEFAULT_BASE_UPLOAD_URL)
+        );
+        String authorizationURL =
+            getKeyValueOrDefault(json, "authorizationURL", DEFAULT_BASE_AUTHORIZATION_URL);
         boolean autoRefresh = json.get("autoRefresh").asBoolean();
 
         // Try to read deprecated value
@@ -865,6 +868,26 @@ public class BoxAPIConnection {
             this.maxRetryAttempts = maxRetryAttempts;
         }
 
+    }
+
+    private String adoptBaseUrlWhenLoadingFromOldVersion(String url) {
+        if (url == null) {
+            return null;
+        }
+        String urlEndingWithSlash = fixBaseUrl(url);
+        return urlEndingWithSlash.equals("https://api.box.com/2.0/")
+            ? DEFAULT_BASE_URL
+            : urlEndingWithSlash;
+    }
+
+    private String adoptUploadBaseUrlWhenLoadingFromOldVersion(String url) {
+        if (url == null) {
+            return null;
+        }
+        String urlEndingWithSlash = fixBaseUrl(url);
+        return urlEndingWithSlash.equals("https://upload.box.com/api/2.0/")
+            ? DEFAULT_BASE_UPLOAD_URL
+            : urlEndingWithSlash;
     }
 
     protected String getKeyValueOrDefault(JsonObject json, String key, String defaultValue) {