+++ src/test/java/com/box/sdk/BoxAPIConnectionTest.java
@@ -19,6 +19,7 @@ import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.startsWith;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
@@ -502,6 +503,126 @@ public class BoxAPIConnectionTest {
         assertThat(api.getBaseUploadURL(), is(restoredApi.getBaseUploadURL()));
         assertThat(api.getAutoRefresh(), is(restoredApi.getAutoRefresh()));
         assertThat(api.getMaxRetryAttempts(), is(restoredApi.getMaxRetryAttempts()));
+    }
+
+    @Test
+    public void restoresProperUrlsFromOldVersions() {
+        // given
+        String accessToken = "access_token";
+        String clientId = "some_client_id";
+        String clientSecret = "some_client_secret";
+        String refreshToken = "some_refresh_token";
+        BoxAPIConnection api = new BoxAPIConnection(clientId, clientSecret, accessToken, refreshToken);
+        api.setBaseURL("https://api.box.com/2.0/");
+        api.setBaseUploadURL("https://upload.box.com/api/2.0/");
+        String savedConnection = api.save();
+
+        // when
+        BoxAPIConnection restoredApi = BoxAPIConnection.restore(clientId, clientSecret, savedConnection);
+
+        // then
+        assertThat("https://api.box.com/2.0/", is(restoredApi.getBaseURL()));
+        assertThat("https://upload.box.com/api/2.0/", is(restoredApi.getBaseUploadURL()));
+    }
+
+    @Test
+    public void restoresProperUrlsWhenNotSet() {
+        // given
+        String accessToken = "access_token";
+        String clientId = "some_client_id";
+        String clientSecret = "some_client_secret";
+        String refreshToken = "some_refresh_token";
+        BoxAPIConnection api = new BoxAPIConnection(clientId, clientSecret, accessToken, refreshToken);
+        api.setBaseURL("https://api.box.com/2.0/");
+        String savedConnection = api.save();
+
+        // when
+        BoxAPIConnection restoredApi = BoxAPIConnection.restore(clientId, clientSecret, savedConnection);
+
+        // then
+        assertThat("https://api.box.com/2.0/", is(restoredApi.getBaseURL()));
+        assertThat("https://upload.box.com/api/2.0/", is(restoredApi.getBaseUploadURL()));
+    }
+
+    @Test
+    public void restoresWithEmptyRefreshToken() {
+        // given
+        String clientId = "some_client_id";
+        String clientSecret = "some_client_secret";
+        String state = "{\"accessToken\":\"some-access-token\","
+            + "\"refreshToken\":null,"
+            + "\"lastRefresh\":1672759898468,"
+            + "\"expires\":3829000,"
+            + "\"userAgent\":\"Box Java SDK v3.8.1 (Java 17.0.4)\","
+            + "\"tokenURL\":null,"
+            + "\"revokeURL\":null,"
+            + "\"baseURL\":\"https://api.box.com/\","
+            + "\"baseUploadURL\":\"https://upload.box.com/api/\","
+            + "\"authorizationURL\":\"https://account.box.com/api/\","
+            + "\"autoRefresh\":true,"
+            + "\"maxRetryAttempts\":5}";
+
+        // when
+        BoxAPIConnection restoredApi = BoxAPIConnection.restore(clientId, clientSecret, state);
+
+        // then
+        assertThat("some-access-token", is(restoredApi.getAccessToken()));
+        assertNull(restoredApi.getRefreshToken());
+        assertThat("https://api.box.com/2.0/", is(restoredApi.getBaseURL()));
+        assertThat("https://upload.box.com/api/2.0/", is(restoredApi.getBaseUploadURL()));
+    }
+
+    @Test
+    public void restoresProperUrlsWhenDeprecatedUrlsAreSet() {
+        // given
+        String clientId = "some_client_id";
+        String clientSecret = "some_client_secret";
+
+        String savedConnection = "{"
+            + "\"accessToken\":\"access_token\","
+            + "\"refreshToken\":\"some_refresh_token\","
+            + "\"lastRefresh\":0,"
+            + "\"expires\":0,"
+            + "\"userAgent\":\"Box Java SDK v3.8.0 (Java 1.8.0_345)\","
+            + "\"tokenURL\":\"https://api.box.com/token\","
+            + "\"revokeURL\":\"https://api.box.com/revoke\","
+            + "\"baseURL\":\"https://api.box.com/\","
+            + "\"baseUploadURL\":\"https://upload.box.com/api/\","
+            + "\"authorizationURL\":\"https://account.box.com/api/\","
+            + "\"autoRefresh\":true,\"maxRetryAttempts\":5"
+            + "}";
+        // when
+        BoxAPIConnection restoredApi = BoxAPIConnection.restore(clientId, clientSecret, savedConnection);
+
+        // then
+        assertThat("https://api.box.com/token", is(restoredApi.getTokenURL()));
+        assertThat("https://api.box.com/revoke", is(restoredApi.getRevokeURL()));
+    }
+
+    @Test
+    public void restoresProperUrlsWhenSomeAreMissing() {
+        // given
+        String accessToken = "access_token";
+        String clientId = "some_client_id";
+        String clientSecret = "some_client_secret";
+        String refreshToken = "some_refresh_token";
+
+        String savedConnection = "{"
+            + "\"accessToken\":\"access_token\","
+            + "\"refreshToken\":\"some_refresh_token\","
+            + "\"lastRefresh\":0,"
+            + "\"expires\":0,"
+            + "\"userAgent\":\"Box Java SDK v3.8.0 (Java 1.8.0_345)\","
+            + "\"baseURL\":\"https://api.box.com/\","
+            + "\"authorizationURL\":\"https://account.box.com/api/\","
+            + "\"autoRefresh\":true,\"maxRetryAttempts\":5"
+            + "}";
+        // when
+        BoxAPIConnection restoredApi = BoxAPIConnection.restore(clientId, clientSecret, savedConnection);
+
+        // then
+        assertThat("https://api.box.com/2.0/", is(restoredApi.getBaseURL()));
+        assertThat("https://upload.box.com/api/2.0/", is(restoredApi.getBaseUploadURL()));
     }
 
     @Test